+++ packages/oauth2-client/src/node/constants.ts
@@ -1,2 +1,4 @@
 export const AUTHORIZATION_REQUEST_BASE_URI = 'malagu.oauth2.client.authorizationRequestBaseUri';
 export const DEFAULT_REDIRECT_URI = 'malagu.oauth2.client.defaultRedirectUri';
+export const AUTHORIZATION_CODE_GRANT_URI = 'malagu.oauth2.client.authorizationCodeGrantUri';
+export const REDIRECT_LOGIN_PATH = 'malagu.oauth2.client.redirectLoginPath';